requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0
schedule==1.2.0
pytz==2023.3
PyYAML==6.0.1
psutil==5.9.6 